refactoring harmonics out to its own subsubmod
